rename service -> repository package
