+++ fixbackend/cloud_accounts/service.py
@@ -24,7 +24,7 @@ from fixbackend.cloud_accounts.models import AwsCloudAccess, CloudAccount
 from fixbackend.cloud_accounts.repository import CloudAccountRepository, CloudAccountRepositoryDependency
 from fixbackend.ids import CloudAccountId, ExternalId, TenantId
 from fixbackend.organizations.dependencies import OrganizationServiceDependency
-from fixbackend.organizations.service import OrganizationService
+from fixbackend.organizations.repository import OrganizationService
 from fixcloudutils.redis.event_stream import RedisStreamPublisher
 from fixbackend.dependencies import FixDependency
