+++ fixbackend/organizations/dependencies.py
@@ -4,7 +4,7 @@ from fastapi import Depends
 
 from fixbackend.db import AsyncSessionDependency
 from fixbackend.dependencies import FixDependency
-from fixbackend.organizations.service import OrganizationService
+from fixbackend.organizations.repository import OrganizationService
 
 
 async def get_organization_service(session: AsyncSessionDependency, fix: FixDependency) -> OrganizationService:
